fixing internal/config/controller/ package
